requirements.txt
langchain>=0.2.0
langchain-groq>=0.1.6
langchain-community>=0.2.0
langgraph>=0.1.0
faiss-cpu>=1.7.4
streamlit>=1.36.0
python-dotenv>=1.0.1
requests>=2.32.3
tiktoken>=0.7.0
